Small inconsistency in updating-arrays-in-state.md fixed
The code snippet in the text is currently referring to a function called "updateMyTodos". However, this function does not exist in the interactive code example. Instead, the code snippet should probably use the function name "updateMyList" (which does exist in the example) to improve the consistency.

This is likely a tiny error.